pyproject.toml
[tool.poetry]
name = "rezipes"
version = "0.1.0"
description = "A recipes sharing web application"
authors = ["Ayomide <ayomidet905@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
djoser = "^2.2.0"
django-rest-framework = "^0.1.0"
rest-framework-simplejwt = "^0.0.2"
django-jazzmin = "^2.6.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
